requirements.txt
aiofiles==23.1.0
aioredis==2.0.1
aiosqlite==0.17.0
anyio==3.7.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==41.0.1
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.99.0
fastapi-admin==1.0.4
fastapi-login==1.9.0
fastapi-restful==0.4.5
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
outcome==1.2.0
passlib==1.7.4
pendulum==2.1.2
psutil==5.9.5
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.10
pyflakes==3.0.1
PyJWT==2.7.0
PyMySQL==1.1.0
pypika-tortoise==0.1.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.17
starlette==0.27.0
tortoise-orm==0.19.3
trio==0.21.0
typing_extensions==4.7.0
urllib3==2.0.3
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
